% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{find_best_lambda}
\alias{find_best_lambda}
\title{Determines the best penalty term from a set of options}
\usage{
find_best_lambda(train_dat, test_dat, response, lambdas)
}
\arguments{
\item{train_dat}{A data frame to construct the model from}

\item{test_dat}{A data frame to test the model on}

\item{response}{The name of a response variable in the data frame (unquoted)}

\item{lambda}{A vector of penalty terms to try}
}
\value{
A data frame of penalty terms and resulting errors
}
\description{
This function uses a randomly chosen test and training set
}
\details{
No interaction terms are included.
}
