% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{find_best_lambda_helper}
\alias{find_best_lambda_helper}
\title{Calculates the SSE using training and test data depending on a given lambda value}
\usage{
find_best_lambda_helper(x, y, y_test, lambda)
}
\arguments{
\item{x}{A matrix of observations and intercepts' column of ones}

\item{y}{A matrix of a response variable's training observations}

\item{y_test}{A matrix of a response variable's test observations}

\item{lambda}{A single lambda value}
}
\value{
A dataframe of the lambda value and its SSE based on test data
}
\description{
This helper function only takes a single lambda value
}
