% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_alt.R
\name{gramsch}
\alias{gramsch}
\title{Implements the Gram-Schmidt process for creating the Q and R matrices}
\usage{
gramsch(x)
}
\arguments{
\item{x}{the data matrix to be decomposed}
}
\value{
a list of the q and r matrices
}
\description{
Implements the Gram-Schmidt process for creating the Q and R matrices
}
