% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_from_coefs.R
\name{predict_from_coefs}
\alias{predict_from_coefs}
\title{Computes predicted values given coefficients}
\usage{
predict_from_coefs(dat, response, coefs)
}
\arguments{
\item{dat}{A data frame}

\item{response}{The name of a response variable in the data frame (unquoted)}

\item{coefs}{A data frame of coefficient estimates}
}
\value{
A data frame of true and predicted values
}
\description{
This function takes a data frame of coefficients in the form outputted by
functions like \code{multiple_linear_regression} or \code{ridge_regression}.
}
\details{
It calculates the predicted values of a response variable from these coefficients.
}
