% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_alt.R
\name{proj}
\alias{proj}
\title{A small function to implement linear projections}
\usage{
proj(u, a)
}
\arguments{
\item{u}{vector to be projected}

\item{a}{vector to be projected upon}
}
\value{
A vector: the projection of u onto a
}
\description{
A small function to implement linear projections
}
