% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Implements ridge regression with many predictors}
\usage{
ridge_regression(dat, response, lambda)
}
\arguments{
\item{dat}{A data frame}

\item{response}{The name of a response variable in the data frame (unquoted)}

\item{lambda}{A vector of penalty terms to try}
}
\value{
A data frame of coefficients
}
\description{
This function computes coefficients for ridge regression
All columns of the provided data frame are used as predictors, except the
one specified as a response.
}
\details{
No interaction terms are included.
}
