% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression_coefs}
\alias{ridge_regression_coefs}
\title{Calculates the matrix of Coefficients used for ridge regression}
\usage{
ridge_regression_coefs(x, y, lambda)
}
\arguments{
\item{x}{A matrix of observations and intercepts' column of ones}

\item{y}{A matrix of a response variable's observations}

\item{lambda}{A single lambda value}
}
\value{
A dataframe of coefficients
}
\description{
This helper function only takes a single lambda value
}
