% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{simple_linear_regression}
\alias{simple_linear_regression}
\title{Implements simple linear regression by hand}
\usage{
simple_linear_regression(dat, response, explanatory)
}
\arguments{
\item{dat}{A data frame}

\item{response}{The name of a response variable in the data frame (unquoted)}

\item{explanatory}{The name of the explanatory variable in the data frame (unquoted)}
}
\value{
A data frame of coefficients
}
\description{
Implements simple linear regression by hand
}
